+++ R/find_countrycol.R
@@ -3,13 +3,12 @@
 #' This function takes a data frame as argument and returns the column name (or index) of all columns containing country names.
 #' For the purpose of this function, a country is any of the 249 territories designated in the ISO standard \code{3166}.
 #' @param x A data frame object
-#' @param return_index A logical value indicating whether the function should return the index of country columns instead of the column names. Default is FALSE.
+#' @param return_index A logical value indicating whether the function should return the index of country columns instead of the column names. Default is \code{FALSE}, column names are returned.
 #' @return Returns a vector of country names (\code{return_index=FALSE}) or column indices (\code{return_index=TRUE}) of columns containing country names.
-#' @seealso \link[countries]{is_country} \link[countries]{country_name}
+#' @seealso \link[countries]{is_country} \link[countries]{country_name} \link[countries]{find_timecol}
 #' @export
 #' @examples
 #' find_countrycol(x=data.frame(a=c("Brésil","Tonga","FRA"), b=c(1,2,3)))
-
 find_countrycol <- function(x, return_index=FALSE){
 
   #check inputs